Further testing
Getting data from database and storing them in file

+++ test.py
@@ -1,8 +1,13 @@
 from scipy import stats
 import mysql.connector
 from mysql.connector import Error
+import os.path
+
+x = []
+y = []
 
 # Test connection to database
+# Test extraction of data
 try:
     db_connection = mysql.connector.connect(host='host',
                                          database='database',
@@ -21,22 +26,32 @@ try:
         # from_edge refers to blocks from top or left end
         create_data_table = 'CREATE TABLE IF NOT EXISTS data (data_id INTEGER(11) AUTO_INCREMENT PRIMARY KEY, line_id INTEGER(11) NOT NULL, succession TINYINT(4) NOT NULL, from_edge TINYINT(4) NOT NULL, block_number TINYINT(4) NOT NULL, longest_block TINYINT(4) NOT NULL, FOREIGN KEY (line_id) REFERENCES line(line_id));'
         db_interaction.execute(create_data_table)
+
+        db_interaction.execute("SELECT line_id FROM line")
+        line_results = db_interaction.fetchall()
+        for line_id in line_results:
+            x.append(line_id[0])
+
+        second_db_interaction = db_connection.cursor()
+        second_db_interaction.execute("SELECT length FROM line")
+        length_results = second_db_interaction.fetchall()
+        for length in length_results:
+            y.append(length[0])
 except Error as error:
     print("Error while connecting to MySQL", error)
 finally:
     if (db_connection.is_connected()):
         db_interaction.close()
+        second_db_interaction.close()
         db_connection.close()
 
 # Test storing data in file and extracting them
-x = [0,2,1,3]
-y = [2,3,4,5]
-
-with open('x_values.txt', 'w') as x_file:
-    for number in x:
-        entry = str(number) + " "
-        x_file.write(entry)
-
+if os.path.isfile('x_values.txt'):
+    with open('x_values.txt', 'w') as x_file:
+        for number in x:
+            entry = str(number) + " "
+            x_file.write(entry)
+"""
 x_read = []
 with open('x_values.txt', 'r') as x_file:
     for line in x_file:
@@ -44,8 +59,10 @@ with open('x_values.txt', 'r') as x_file:
         words = line.split(' ')
         for number in words:
             x_read.append(int(number))
+"""
 
 # Test linear regression analysis
-slope, intercept, r_value, p_value, std_err = stats.linregress(x_read,y)
+slope, intercept, r_value, p_value, std_err = stats.linregress(x,y)
 
 print ("R-squared:", r_value**2)
+print ("P-value:", p_value)
